+++ Space-Invaders/Header/Player/PlayerView.h
@@ -1,31 +1,33 @@
 #pragma once
 #include <SFML/Graphics.hpp>
 
-class PlayerController;
+namespace Player {
+    class PlayerController;
 
-class PlayerView
-{
-private:
+    class PlayerView
+    {
+    private:
 
-    const sf::String player_texture_path = "assets/textures/player_ship.png";
-    const float player_sprite_width = 60.f;
-    const float player_sprite_height = 60.f;
+        const sf::String player_texture_path = "assets/textures/player_ship.png";
+        const float player_sprite_width = 60.f;
+        const float player_sprite_height = 60.f;
 
-    sf::RenderWindow* game_window;
+        sf::RenderWindow* game_window;
 
-    sf::Texture player_texture;
-    sf::Sprite player_sprite;
+        sf::Texture player_texture;
+        sf::Sprite player_sprite;
 
-    void initializePlayerSprite();
-    void scalePlayerSprite();
+        void initializePlayerSprite();
+        void scalePlayerSprite();
 
-    PlayerController* player_controller;
+        PlayerController* player_controller;
 
-public:
-    PlayerView();
-    ~PlayerView();
+    public:
+        PlayerView();
+        ~PlayerView();
 
-    void initialize(PlayerController* controller);
-    void update();
-    void render();
-};
+        void initialize(PlayerController* controller);
+        void update();
+        void render();
+    };
+}
