+++ Space-Invaders/Header/PlayerService.h
@@ -2,37 +2,40 @@
 #include <SFML/Graphics.hpp>
 #include "../Header/Player/PlayerController.h"
 
-class PlayerService
-{
 
-private:
-    
-    int health = 3;
-    sf::Vector2f position = sf::Vector2f(200.0f, 100.0f);
-    int movement_speed = 5;
-    int player_score = 0;
+namespace Player {
+    class PlayerService
+    {
 
-    const sf::String player_texture_path = "assets/textures/player_ship.png";
+    private:
 
-    sf::Texture player_texture;
-    sf::Sprite player_sprite;
+        int health = 3;
+        sf::Vector2f position = sf::Vector2f(200.0f, 100.0f);
+        int movement_speed = 5;
+        int player_score = 0;
 
-    sf::RenderWindow* game_window; //as always
+        const sf::String player_texture_path = "assets/textures/player_ship.png";
 
-    void initializePlayerSprite();
-    void processPlayerInput();
+        sf::Texture player_texture;
+        sf::Sprite player_sprite;
 
-public:
+        sf::RenderWindow* game_window; //as always
 
-    PlayerService();
-    ~PlayerService();
+        void initializePlayerSprite();
+        void processPlayerInput();
 
-    void initialize();
-    void update();
-    void render();
+    public:
 
-    void move(float offsetX);
-    int getMoveSpeed();
-    sf::Vector2f getPlayerPosition();
+        PlayerService();
+        ~PlayerService();
 
-};
+        void initialize();
+        void update();
+        void render();
+
+        void move(float offsetX);
+        int getMoveSpeed();
+        sf::Vector2f getPlayerPosition();
+
+    };
+}
