+++ Space-Invaders/Source/Player/PlayerModel.cpp
@@ -1,45 +1,47 @@
 #include "../../Header/Player/PlayerModel.h"
 
-PlayerModel::PlayerModel() { }
+namespace Player {
+	PlayerModel::PlayerModel() { }
 
-PlayerModel::~PlayerModel() { }
+	PlayerModel::~PlayerModel() { }
 
-void PlayerModel::initialize() { reset(); } // remember to call reset() 
+	void PlayerModel::initialize() { reset(); } // remember to call reset() 
 
-void PlayerModel::reset()
-{
-	player_state = PlayerState::ALIVE; // set state to alive
-	player_position = initial_player_position;
-	player_score = 0;
-}
+	void PlayerModel::reset()
+	{
+		player_state = PlayerState::ALIVE; // set state to alive
+		player_position = initial_player_position;
+		player_score = 0;
+	}
 
-sf::Vector2f PlayerModel::getPlayerPosition()
-{
-	return player_position;
-}
+	sf::Vector2f PlayerModel::getPlayerPosition()
+	{
+		return player_position;
+	}
 
-void PlayerModel::setPlayerPosition(sf::Vector2f position)
-{
-	player_position = position;
-}
+	void PlayerModel::setPlayerPosition(sf::Vector2f position)
+	{
+		player_position = position;
+	}
 
-int PlayerModel::getPlayerScore()
-{
-	return player_score;
-}
+	int PlayerModel::getPlayerScore()
+	{
+		return player_score;
+	}
 
-void PlayerModel::setPlayerScore(int score)
-{
-	player_score = score;
-}
+	void PlayerModel::setPlayerScore(int score)
+	{
+		player_score = score;
+	}
 
-//..
-PlayerState PlayerModel::getPlayerState()
-{
-	return player_state;
-}
+	//..
+	PlayerState PlayerModel::getPlayerState()
+	{
+		return player_state;
+	}
 
-void PlayerModel::setPlayerState(PlayerState state)
-{
-	player_state = state;
+	void PlayerModel::setPlayerState(PlayerState state)
+	{
+		player_state = state;
+	}
 }
