+++ Space-Invaders/Space-Invaders.vcxproj
@@ -133,25 +133,25 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="Source\EventService.cpp" />
-    <ClCompile Include="Source\GameService.cpp" />
-    <ClCompile Include="Source\GraphicService.cpp" />
+    <ClCompile Include="Source\Event\EventService.cpp" />
+    <ClCompile Include="Source\Gameplay\GameService.cpp" />
+    <ClCompile Include="Source\Graphic\GraphicService.cpp" />
     <ClCompile Include="Source\Player\PlayerController.cpp" />
     <ClCompile Include="Source\Player\PlayerModel.cpp" />
     <ClCompile Include="Source\PlayerService.cpp" />
     <ClCompile Include="Source\Player\PlayerView.cpp" />
-    <ClCompile Include="Source\ServiceLocator.cpp" />
+    <ClCompile Include="Source\Global\ServiceLocator.cpp" />
     <ClCompile Include="Source\TimeService.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Header\EventService.h" />
-    <ClInclude Include="Header\GameService.h" />
-    <ClInclude Include="Header\GraphicService.h" />
+    <ClInclude Include="Header\Event\EventService.h" />
+    <ClInclude Include="Header\Gameplay\GameService.h" />
+    <ClInclude Include="Header\Graphic\GraphicService.h" />
     <ClInclude Include="Header\Player\PlayerController.h" />
     <ClInclude Include="Header\Player\PlayerModel.h" />
     <ClInclude Include="Header\PlayerService.h" />
     <ClInclude Include="Header\Player\PlayerView.h" />
-    <ClInclude Include="Header\ServiceLocator.h" />
+    <ClInclude Include="Header\Global\ServiceLocator.h" />
     <ClInclude Include="Header\TimeService.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
