+++ Space-Invaders/Space-Invaders.vcxproj.filters
@@ -18,16 +18,16 @@
     <ClCompile Include="main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Source\GameService.cpp">
+    <ClCompile Include="Source\Gameplay\GameService.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Source\ServiceLocator.cpp">
+    <ClCompile Include="Source\Global\ServiceLocator.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Source\GraphicService.cpp">
+    <ClCompile Include="Source\Graphic\GraphicService.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Source\EventService.cpp">
+    <ClCompile Include="Source\Event\EventService.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="Source\PlayerService.cpp">
@@ -47,16 +47,16 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Header\GameService.h">
+    <ClInclude Include="Header\Gameplay\GameService.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Header\ServiceLocator.h">
+    <ClInclude Include="Header\Global\ServiceLocator.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Header\GraphicService.h">
+    <ClInclude Include="Header\Graphic\GraphicService.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="Header\EventService.h">
+    <ClInclude Include="Header\Event\EventService.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Header\PlayerService.h">
